backend/requirements.txt
Celery==4.4.6
Click==7.0
Cryptography==2.9.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Gzip==0.2
Flask-Login==0.5.0
Flask-Marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
GitPython==3.1.2
Marshmallow==3.5.1
Marshmallow-Enum==1.5.1
Marshmallow-Sqlalchemy==0.22.3
Pandas==1.0.3
Python-Box==4.2.2
PyJWT==1.7.1
PyOdbc==4.0.30
PsUtil==5.7.2
PyWin32==223
Requests==2.23.0
Waitress==1.4.4
Xlwt==1.3.0
Xlrd==1.2.0
